requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
neo4j==4.4.5
python-dotenv==1.0.0
psycopg2-binary==2.9.5
